% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow-registry.R
\name{.flow_metric_registry}
\alias{.flow_metric_registry}
\title{Build (or retrieve) flow metric registry}
\usage{
.flow_metric_registry()
}
\value{
Tibble with flow metric metadata and runtime fields.
}
\description{
Build (or retrieve) flow metric registry
}
\keyword{internal}
