% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow-registry.R
\name{.flow_registry_invalidate}
\alias{.flow_registry_invalidate}
\title{Invalidate cached flow registry objects}
\usage{
.flow_registry_invalidate()
}
\value{
Invisible \code{TRUE}.
}
\description{
Invalidate cached flow registry objects
}
\keyword{internal}
