% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eflow_metrics.R
\name{.hydro_metric_registry_base}
\alias{.hydro_metric_registry_base}
\title{Build (or retrieve from cache) the base metric registry.}
\usage{
.hydro_metric_registry_base()
}
\value{
Tibble with metric definitions.
}
\description{
Build (or retrieve from cache) the base metric registry.
}
\keyword{internal}
