% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eflow_metrics.R
\name{.hydro_metric_registry_base_build}
\alias{.hydro_metric_registry_base_build}
\title{Internal: construct the base metric registry from scratch.}
\usage{
.hydro_metric_registry_base_build()
}
\value{
Tibble with metric definitions including grain and purpose list-column.
}
\description{
Internal: construct the base metric registry from scratch.
}
\keyword{internal}
