% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydro_metrics.R
\name{.hydro_metric_sets}
\alias{.hydro_metric_sets}
\title{Return the named list of metric sets used for selection.}
\usage{
.hydro_metric_sets()
}
\value{
Named list of character vectors of metric keys.
}
\description{
Return the named list of metric sets used for selection.
}
\keyword{internal}
