% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow-calculate.R
\name{.hydro_rollmean_right}
\alias{.hydro_rollmean_right}
\title{Right-aligned rolling mean with NA padding.}
\usage{
.hydro_rollmean_right(x, k)
}
\description{
Right-aligned rolling mean with NA padding.
}
\keyword{internal}
