% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow-calculate.R
\name{.hydro_smooth_center}
\alias{.hydro_smooth_center}
\title{Centered rolling mean with NA padding.}
\usage{
.hydro_smooth_center(x, k)
}
\description{
Centered rolling mean with NA padding.
}
\keyword{internal}
