% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.is_chr_scalar}
\alias{.is_chr_scalar}
\title{Check if value is a non-empty character scalar}
\usage{
.is_chr_scalar(x)
}
\arguments{
\item{x}{Value to check.}
}
\value{
Logical.
}
\description{
Check if value is a non-empty character scalar
}
\keyword{internal}
