% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.is_int_scalar}
\alias{.is_int_scalar}
\title{Check if value is a scalar integer}
\usage{
.is_int_scalar(x)
}
\arguments{
\item{x}{Value to check.}
}
\value{
Logical.
}
\description{
Check if value is a scalar integer
}
\keyword{internal}
