% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.is_num_scalar}
\alias{.is_num_scalar}
\title{Check if value is a scalar numeric}
\usage{
.is_num_scalar(x)
}
\arguments{
\item{x}{Value to check.}
}
\value{
Logical.
}
\description{
Check if value is a scalar numeric
}
\keyword{internal}
