% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drought-catalog.R
\name{.metric_catalog_drought}
\alias{.metric_catalog_drought}
\title{Internal drought metric catalog}
\usage{
.metric_catalog_drought()
}
\value{
Tibble with metadata columns:
key, family, group, grain, label, input.
}
\description{
Internal drought metric catalog
}
\keyword{internal}
