% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wr-catalog.R
\name{.metric_catalog_water_resources}
\alias{.metric_catalog_water_resources}
\title{Internal water-resources metric catalog}
\usage{
.metric_catalog_water_resources()
}
\value{
Tibble with metadata columns:
key, family, group, grain, label, input.
}
\description{
Internal water-resources metric catalog
}
\keyword{internal}
