% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/erfa-metric-set.R
\name{.metric_set_erfa70}
\alias{.metric_set_erfa70}
\title{ERFA70 metric-set descriptor}
\usage{
.metric_set_erfa70()
}
\value{
A list with components:
name, keys, and metadata (key/group/grain/label).
}
\description{
ERFA70 metric-set descriptor
}
\keyword{internal}
