% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eflow_erfa.R
\name{.print_erfa_block}
\alias{.print_erfa_block}
\title{Internal: print a single ERFA summary block}
\usage{
.print_erfa_block(s, combo_row, extra_cols)
}
\arguments{
\item{s}{Summary tibble.}

\item{combo_row}{One-row data.frame of grouping values, or NULL.}

\item{extra_cols}{Character vector of grouping column names.}
}
\description{
Internal: print a single ERFA summary block
}
\keyword{internal}
