% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydrokit-package.R
\docType{package}
\name{hydrokit-package}
\alias{hydrokit}
\alias{hydrokit-package}
\title{hydrokit: Hydrological Indicators for Climate and Water Resource Analysis}
\description{
Collection of R scripts for processing, analysis, and visualization of hydrological data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/tanerumit/hydrokit}
  \item Report bugs at \url{https://github.com/tanerumit/hydrokit/issues}
}

}
\author{
\strong{Maintainer}: Umit Taner \email{umit.taner@deltares.nl}

}
\keyword{internal}
